requirements.txt
cbbd>=1.23.0
python-dotenv>=1.0.0
scipy>=1.12.0
numpy>=1.26.0
